backend/requirements.txt
Flask
Flask-Cors
psycopg2-binary
Flask-SQLAlchemy
Flask-Migrate
Authlib
Flask-Bcrypt
pillow